0810_c2 #detail 페이지 top부분 스크립트 진행 중 (90%)
